KafkaScripts/requirements.txt
# Install with: pip install -r requirements.txt
kafka-python==2.0.2
pandas==2.2.2
requests==2.32.3
